Add species page, and reorganize file/folder structure
