+++ views/navigation.ejs
@@ -15,10 +15,16 @@
     <div class="collapse navbar-collapse ms-2" id="navbarNavAltMarkup">
       <div class="navbar-nav me-auto">
         <a
-          class="nav-link<%= (page === 'new') ? '' : ' active' %>"
-          <%= (page === 'new') ? '' : 'aria-current="page"' %>
+          class="nav-link<%= (page === 'new') || (page === 'species') ? '' : ' active' %>"
+          <%= (page === 'new') || (page === 'species') ? '' : 'aria-current="page"' %>
           href="/"
           >Home</a
+        >
+        <a
+          class="nav-link<%= (page === 'species') ? ' active' : '' %>"
+          <%= (page === 'species') ? 'aria-current="page"' : '' %>
+          href="/species/all"
+          >Species</a
         >
         <% if (user) { %>
         <a
